Modified Absorber to only fire when path is clear.
This involved modifying several other classes - the relevant code was
moved from view.Board to model.GameModel, and the absorber's
constructors had to be modified to include a reference to the model.
Model now contains methods getGizmo(Vect coords) and isCellEmpty(Vect
coords), which check all gizmos to see if they occupy that square.