Remove double-triggering of absorbers on keystroke.
This is a nasty hack - key triggers are added in RunKeyListener, when
they should be handled in AddKeyTriggerListener. RunKeyListener is the
key listener for run mode. For now I've just put in a check for if a
gizmo is a flipper, and only add a keyPressedTrigger if it is. Probably
most of the actionListeners will have to be rewritten to handle their
own use cases rather than pass it off to the BoardMouseListener.